moved followTrajectory to drive, removed logging from routines
